Pathologies API refactoring.
Converted all DTOs to records, except from CommonDataElement because we need to modify it.
Created a PathologyService and extracted all complexity from the Controller.
Renamed the MetadataHierarchyDTO to PathologyMetadataDTO.
Created a new PathologyMetadataGroupDTO class for the metadata groups. The PathologyMetadataDTO class has more properties now that were not included previously.
Removed the Exareme2AttributesDTO, getting the internal info hardcoded, maybe we should change that in the future.